Recursive find for binary tree
This turns out to be less trivial than I thought. After spending
a couple of days banging on it, this commit represents a more
systematic approach using more tests with shallower trees. The
idea is getting it right for depth 1 and 2 trees, then seeing if
that works for trees of arbitrary depth.