first cut at breadth first search
Not really happy with this implementation, and it's tested
only on a shallow, balanced tree. There should be a way to
do it in one function, with a flat array instead of nested
arrays.

I don't see any way to do it without using at least two arrays
though, one to keep a list of the current nodes, another to
keep track of where the nodes are.