ruby - bad code for successor
Here's a bunch of really bad code for finding successors in
binary search trees. I worked on this for a very long day
thinking I had it nailed, then found an oversight. The code
in this commit is an attempt to rectify the oversight. I ended
up coding myself into a hole, and looking up an algorithm for
successor in CLR. It turns out I had it almost completely
correct the first time!

Some of the code in this commit may be salvageable.

Note the successor algorithm implemented here does not depend
on having each node linked to its parent.