ruby - delete from CLRS3
Apparently I now have a working delete function at the tree
level for ruby, taken from CLRS 3rd Edition.

The code is not Ruby-ish, but it does pass several hundreds lines
of testing.

A next goal for the delete subproject is tasking out the remainder
of the `transplant` implementations, and figuring out how to clean
up the CLRS code into something a bit more literate. The literate
part should be pretty easy in Ruby, not sure what can be done in
other languages.

Another worth goal here would be finding a way to more easily
generate test data such that alternate implementations can be
tested. That's probably a lot of work, not sure what the return
would, other then deeper understanding, which may be worth it
in any case.