c - start of transplant implementation
**This commit doesn't pass testing**

Implementing transplant is easy. However, continually
rebuilding trees from scratch is boring, hence an `unlink`
function will allow rebuilding trees from arbitrary nodes.

The unlink function in turn depends on the post_order_traverse
function, which depends on an out-of-date technique for
implementing the relevant closure.

At the moment, fixing the closure is the next item of work.
That will be followed by finishing the unlink implementation
(which also needs to be tested), then transplant.