c - transplant, using delete in testing
After an attempt to unlink the trees built for
testing transplant, it became apparent that it's
unreasonable to assume a tree with transplanted
nodes could be guaranteed to unlink.

After implementing delete in its place, it became
apparent that assuming delete would collect all
the memory correctly is also a non-starter. However,
these tests are small enough that a future iteration
can simply free the relevant trees and nodes directly.