c++ - transplant node implemented for clrs delete
This implementation of transplant went very quickly
as all the tree and node instantiation code was moved
into the spec, rather than trying to deal with somehow
reusing already declared variables in a tree which has
been tampered with (i.e., nodes transplanted).

It's worth noting here, again, that `transplant` would
be better names `reset_parent` as it's limited to managing
only the parent links.