c++ - list_keys and Node::collect
The list_keys method has been discussed elsewhere.

This commit also moved the Tree::collect method into
Node class, something which was left over from the very
early days, and overdue. The idea is to push everything
possible to as low as it can go, and nodes can collect
themselves.