c - collector enhancement
The Collector class gets a rudimentary enhancement to
build from an array of ints. This code isn't complete
as I decided I didn't really need it to finish the c
tree node deletion implementation, and it would be
too time consuming to continue.

The next step would be to thoroughly test it.

Also, a stub for clearing was added, which also
needs to be vetted and tested.