java - node unlink and is_unlinked
unlink and is_unlinked are primarily helper methods for
implementing tree.delete(). The idea is that a deleted
node should not have links to any "live" nodes, it should
just be a piece of unreferenced memory which isn't
referencing anything.

It would be interesting to whether garbage collection
is influenced by deliberating unlinking nodes.