doc - more notes on delete from bst
Delete had given me a bit of a fit. It's not been
a problem to just punch in the algorithm. And the
algorith I devised seem work ok as well, but I've
had trouble getting it into my memory with deep
understanding.

Today's addition is a list of cases for delete, which
seem to fit into my comprehension pretty well. At
some point in the future, I'll recode the delete
algorithms from memory.