hash tree - super dumb document insertion
Get some documents into the tree and hash them at
will.

This commit looks really stupid because it is. Mostly,
it's a way to plow through my current log jam on hash
trees. Not feeling the love for it at the moment, so
just writing anything is a step forward.