ruby tree - iterative thrashing
For whatever reason, my head isn't into working through
iterative traversal at the moment. First, it feels highly
unnatural compared to recursion, and second, I'm just
blocked from it somehow, just don't feel the compelling
need to put the required hour or two of intensive focus
on it. Which is stupid because this is simple. I could
look it up in CLRS, which is a horrible temptation.