sentence numbering macro to bst
I've been looking around for a way to number individual
sentences, and this commit is one, pretty easy way to do
it. I was actually looking for some way to automatically
count sentences, but on further thought, that's a really
hard problem, and not sure it's solveable in TeX. (Or at
all.)

The macro here will probably get moved to some place it's
really needed.