iterative-c - full right subtree iterates
Big news for this commit is I found a really stupid rookie
mistake on array sizing in the collector. As it turns out,
this function was never tested, so it's been working by
accident with the size 2 arrays I've been feeding it. Not
sure it's worth making this function work correctly or not.
Probably so just to ensure that I don't forget how to do this
stuff in c.