bst doc - start double factorial discussion
The double factorial n!! for odd n appears in the derivation
of number of binary trees as a simplification following
Taylor series or binomial expansion of the opsgf. The
discussion in this commit is very crude, it's purpose is
mostly to get the LaTeX punched in correctly. Better prose
will follow in a succeeding commit.

A few notes on the purpose of pre- and post-order commits
also appear.