ruby - pre- and post- order key collect
Because the traversing is written to accept callbacks,
it's trivial to implement a collection on both pre-
and post-order traversing. The reason this was done
is to support pre- and post-order traversing by iteration,
where the test data from the recursive implementation is
used as "known-good" values.

A few other cleanups on the specs to better follow guidelines
promoted by betterspecs.com. This is ongoing.