ruby-pre-iter - implementation complete
The pre-order and in-order traversals are mechanically
very similar, differing mostly in where the values are
printed out. It will be interesting to see how this works
with post-order traversal. I'm going to start the
post-order implementation by assuming the underlying
mechanism is the same as pre-order, and work from that.