ruby-post-iter - horrible commit, not working
This commit needs to be amended, I do not want it
merged into master. At the moment, I'm completely
stuck and want to work on associated material, namely
the doc branch to finish the current refactoring in
the documents.

That said, I think I'm closing in on a viable strategy.
The idea is that leaf nodes do not go onto the stack,
and once a node is written out, it gets marked as visited.
This should allow working up the tree, and should be
amenable to deriving with pencil and paper.