ruby-post-iter - iterate to find unvisited leaf
The main stumbling block I've been having with post-order
iteration is revisiting nodes I've already seen. This puts
the iteration into an endless loop of course. One way to
mitigate is marking nodes seen once the algorithm has
processed appropriately at that node.

To make it easier to monitor what's happening with the
processing, the iteration method(s) need to take blocks
just as the recursive methods do. This is something for
an upcoming commit.