ruby-post-iter - first working version
Remarkably but unsurprisingly simple. Two key notions:

1. visited nodes must be marked so that the procedure
can "climb" back out of the tree;
2. always find the "deepest" leaf at every iteration.

This took way, way too long, as can be seen from the
commit history. There is a temptation to squash this
entire branch, and if this were a production code base,
it would be squashed into a single commit. Since the main
purpose of the entire project is pedagogical, the commit
history remains in place as a reminder of the false
starts and dead ends encountered while learning. I may
not continue this practice in the future, but it stands
for now.