ruby-insert-iter - now actually working
Kind of clunky looking, but that's how it is for
most iterative algorithms. It may be possible to
use ternaries instead of if-else, not sure atm if
Ruby will support that. Something for a followon
commit.

Supposedly, some huge Antarctic ice sheet is about to
collapse, and will raise sea level about 6 feet in a
couple of years. I think I'm ok where I'm at right now.
At 10 feet, probably not.

"This!" she proclaimed, basking in the glow of
delusional conviction acquired from the latest
currently fashionable pop culture aphorism.