clrs - exercise 12.1-3 iterative walk
Provided algorithm based on stack. Not as clean
as I would like it. Will likely revisit this
in the future.

I'm supposed to go to this Elixir meetup tonight,
which is down by Mission Cliff. That's a haul
from the Bart station. At least I'll get my walk in
today. I don't feel like going right now, but I'd
best just suck it up and leave the house when it's
time, just don't even think about it otherwise.

Time to make a burger.