clrs12.1-5 worst case comparisons on arbitrary data
I found this exercise to be difficult for at least
a couple of reasons:

1. Since I haven't worked through all the preliminary
material, I had to backtrack through the text to extract
necessary notions and definitions.

2. Worst-case bounding n lg n doesn't make any sense to
me, because arbitrary data can include a sorted list,
which, how I understand, is going to require n^2 operations
for constructing a tree, unless the list is split by
midpoints to build the tree.

The work stands for now, but I plan on revisiting and
updating in the future, as I understand the math and
conventions for these sorts of arguments.