stern-brocot - rational numbers <= 1
This is partially working, committed to keep
track of where the implementation needs to be
fixed. Whatever the problem is, it's almost
surely very simple.

Medium coffee today, on the start of the ramp down.
Will get a small tomorrow. This is going to work, as
I'm between 1/3 and 1/2 left on the medium and it's
noon already.

Apparently, El Jefe Naranja is losing his marbles
over the Comey thing. Would be very entertaining
except for the nuclear weapons thingy.