doc: cleanup and expository text
I've decided to renew my vow to "ship" documentation
and papers like software: no matter how much or how
little is done, every commit on master should be
shippable as-is.

This is much, much more difficult than it would seem,
especially with years of bad habits. This commit
doesn't make the cut, but it gets a lot closer, and
I need to rebase under a branch for a specific problem,
right now.

For some very odd reason, I was under the impression
that the Cowboy Junkies were from Houston, but they're
not, they Canadian. My impression probably comes from
the TVZ reference on Towne's Blues from Black Eyed Man.
Hrmph.