rspec shared examples for node
This is a task I've punted on several times in the past.
Reading the documentation on shared examples finds that
my use case here isn't quite the use case others have in
mind. Typically, shared examples seem to provide a sort
of "integration" capability, where the shared code is
extracted into a module, then the classes including the
module are tested with respect to the modules behavior.

For modules, I always wrap them in a class and test them
directly, but this is certainly a viable alternative.

What interest in shared examples is to refactor and dry
up tests. The structure and semantics of shared examples
doesn't quite match what I'm looking for: different
classes using the same tests, on per-method basis.

Since this is my first crack at shared examples, it's
likely that I'm missing one or more important concepts.
I plan on using shared examples much more in the future,
so I'm looking forward to getting more out of them.

Meantime, what's been done here is really useful.