postorder walk shared example
Working now for recursive and iterative trees.

There is some mildy cool callbackering going on
the shared example, and the iterative post order
method now takes  callback.

The inorder and preorder are going to take a bit of
time. Nothing hard about either, but there is a fair
bit of code to shuffle around.