shared-walk: removed block arguments
The blocks do not need to be passed around as
arguments, instead can be procified and passed
along. I suppose if the block needed to go more
than a couple of levels, it would be better to
pass it rather than continuing to build new Proc
objects, but in that case, needing to pass so
deeply is its own code smell.