documentation updates
Mostly rearranging and restructuring. I'd like
to get the documentation under regular development
going forward, it has been languishing.

Also added some helper functions for GCD and
coprime, and a stub class for CalkinWilf tree.
These will likely get moved in the future, but
this is as good a place as any to get started.