avl: initial work on balance and weight
The weight was initially computed during insert,
but that's incorrect, it needs to account for the
final configuration of the tree. I think this is
correct now.

There is an overhead associated with computing the
height on demand. I hesitate to memoize it as the
tree will change height at any node over its
lifespan.