inverse: reverse left and right children
This cute little thing comes from an internet rant
posted by the main home brew dude as a result of his
experience interviewing at google. He was unhappy
about being asked to "invert a binary tree."

After reading the rant and a quick google search
to find out what a binary tree inverse was, I spent
maybe 2 minutes cogitation to figure out the problem
wasn't very difficult, just read all the node in
preorder walk and use an insert method which reverses
the inequality. Since that felt tedious and boring,
I decided I'd just punt on my own implementation.

Also, the explanation I found on the web immediately
suggests that "inverse" is the wrong notion. A better
notion would be reverse because the tree can be constructed
by reversing the inequality. Inverse is just the wrong
word for the explanation I found. Perhaps there is
a different explanation, perhaps corresponding to the
rant author's question.

In any case, when I woke up the next morning I knew
the problem was beyond trivial: just swap the left
and right pointers in a preorder traverse. I decided
to implement.

The hardest part of the implementation was figuring out
the appropriate test cases and typing them in. I didn't
do it all in one sitting, just grabbed a few moments
here and there to punch in a few tests. Once I had the
tests mostly complete, I wrote this commit in a few
minutes.

The actual code to do the work is less than 1 minute
of typing, because I've already done the heavy lifting
for the tree walks, all of which take a block. So all
it required was defining an `invert` method, and calling
the preorder walk with a block swapping child pointers.
Done.

In production code, I would squash all the commits on
this branch. For demonstration and learning purposes,
I'm keeping the commits as-is, so that the evolution
of the implementation is recorded for future reference.