avl: more example specs
Along with adding more examples, there is a marker
for working through the mechanics of the recursion
on insert. It's pretty interesting, and not exactly
how I though it would work. Could be as much as a
day's work to go through how the stack works with
respect to arguments passed into the recursing
method, which is overloaded, btw, we're calling super,
that's the interesting part.