semantic adjust to ruby tree transplant
The transplant method used to reset the parent
during node deletion got some variable renaming
to help remember what it's doing. The renaming
isn't completely accurate in the case where the
reset node is not a child of the deleted node.

Also, added a few tests to the gcd routine.