avl: "knee" rotation case from tree retracing
I think this is the commit which finishes the tree
insertion which handles retracing and rebalancing
correctly.

The main thing I've gotten out of this is that the
references I've been using make it more difficult
than it really is, primarily because they all use
large functions which do too much.

There is also the notion of returning an updated
root node from a rotation. This may be a useful notion,
and worth investigating in more detail.

What's next is another round of refactoring the
insertion and rebalancing, then deletions, which I
believe will be far more difficult.